backend/requirements.txt
torch
flask
numpy
datasets
sentence-transformers
faiss-cpu
requests
beautifulsoup4
scikit-learn
nltk
sentencepiece
transformers
rank_bm25
accelerate
sympy
spacy
pillow
pytesseract
